Make CodeQL query more realisitic
Change the CodeQL query to
(a) make the query proper more realistically represent a SQL injection
query;
(b) provide an additional query predicate that exports the
intra-procedural taint flow edges.

The intra-procedural query predicate is designed to be completely
general, so that it can be dropped into any taint analysis without any
knowledge of the details of this analysis.